XW-1104 | renamed remaining mediawiki directories to wiki
